editing touchunlock.fish

+++ touchunlock.fish
@@ -1,8 +1,10 @@
 function touchunlock
     pcsc_scan | while read -l line
-        echo $line
-        if locked?
-            killall i3lock
+        if echo $line | grep '^ccccc'
+            echo read $line
+            if locked?
+                killall i3lock
+            end
         end
     end
 end
